[refactor] Fix paths to import utils
