+++ audyn/models/ast.py
@@ -12,7 +12,7 @@ from omegaconf import OmegaConf
 from torch.nn.common_types import _size_2_t
 
 from ..modules.vit import PositionalPatchEmbedding
-from ..utils.github import download_file_from_github_release
+from ..utils._github import download_file_from_github_release
 
 __all__ = [
     "AudioSpectrogramTransformer",
@@ -406,7 +406,7 @@ class AudioSpectrogramTransformer(BaseAudioSpectrogramTransformer):
                 - ast-base-stride10
 
         """  # noqa: E501
-        from ..utils._hydra._hydra import instantiate  # to avoid circular import
+        from ..utils._hydra.utils import instantiate  # to avoid circular import
 
         pretrained_model_configs = _create_pretrained_model_configs()
