+++ audyn/models/ssast.py
@@ -14,7 +14,7 @@ from omegaconf import OmegaConf
 from torch.nn.common_types import _size_2_t
 
 from ..modules.vit import PositionalPatchEmbedding
-from ..utils.github import download_file_from_github_release
+from ..utils._github import download_file_from_github_release
 from .ast import (
     Aggregator,
     BaseAudioSpectrogramTransformer,
